# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: eshiritori API
  version: 0.0.0
paths:
  /rooms:
    post:
      operationId: createRoom
      summary: 部屋を作る
      description: 作った人がホストで、ただ 1 人のメンバーになった部屋を作る。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "201":
          description: 部屋を作った。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRoomResponse"
        "400":
          description: 入力の誤り、または要求の形の誤り。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 予期しない失敗。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rooms/{roomCode}/members:
    post:
      operationId: joinRoom
      summary: 部屋に入る
      description: 部屋コードの部屋に、ニックネームでメンバーとして加わる。
      parameters:
        - $ref: "#/components/parameters/RoomCode"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRoomRequest"
      responses:
        "201":
          description: 部屋に加わった。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRoomResponse"
        "400":
          description: 入力の誤り、または要求の形の誤り。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 部屋が無い、または期限切れ（room.room_not_found）。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 満員（room.room_full）、ゲームの開始済み（room.game_already_started）、ニックネームの重複（room.nickname_taken）。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 予期しない失敗。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rooms/{roomCode}:
    get:
      operationId: getRoom
      summary: 部屋の状態を取る
      description: 部屋のメンバーだけが取れる。保存はせず、部屋の期限を延ばさない。
      security:
        - playerToken: []
      parameters:
        - $ref: "#/components/parameters/RoomCode"
      responses:
        "200":
          description: 部屋の状態。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetRoomResponse"
        "403":
          description: 部屋のメンバーでない（room.not_member）。トークンが無い・形が違う場合も同じ。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 部屋が無い、または期限切れ（room.room_not_found）。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 予期しない失敗。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rooms/{roomCode}/start:
    post:
      operationId: startGame
      summary: ゲームを始める
      description: ホストだけが始められる。描く順番をランダムに決め、members をその並びにする。既に始まった部屋への要求は何もせず、いまの部屋を 200 で返す（描く順番は決め直さない）。
      security:
        - playerToken: []
      parameters:
        - $ref: "#/components/parameters/RoomCode"
      responses:
        "200":
          description: ゲームが始まった部屋。members が描く順番。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartGameResponse"
        "403":
          description: ホスト以外のメンバー、またはメンバーでない人からの要求（room.not_host）。トークンが無い・形が違う場合も同じ。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 部屋が無い、または期限切れ（room.room_not_found）。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: メンバーがホスト 1 人だけ（room.not_enough_members）。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 予期しない失敗。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    playerToken:
      type: http
      scheme: bearer
      description: 部屋を作る・入る応答の player.playerToken。Authorization ヘッダーに Bearer で載せる。
  parameters:
    RoomCode:
      name: roomCode
      in: path
      required: true
      schema:
        type: string
      description: 部屋コード。大文字小文字・全角半角・前後の空白の違いはそろえて扱う。書式の誤りは room.room_not_found。
  schemas:
    CreateRoomRequest:
      type: object
      required: [nickname, roundCount]
      properties:
        nickname:
          type: string
          description: ホストのニックネーム。前後の空白を除いて 1〜10 文字（絵文字 1 つを 1 文字と数える）。範囲外は room.nickname_empty または room.nickname_too_long の 400 で返す。
        roundCount:
          type: integer
          description: 周回数。1〜5 の整数。範囲外は room.round_count_out_of_range の 400 で返す。
    CreateRoomResponse:
      type: object
      required: [room, player]
      properties:
        room:
          $ref: "#/components/schemas/RoomSnapshot"
        player:
          $ref: "#/components/schemas/PlayerCredential"
    JoinRoomRequest:
      type: object
      required: [nickname]
      properties:
        nickname:
          type: string
          description: 入る人のニックネーム。前後の空白を除いて 1〜10 文字（絵文字 1 つを 1 文字と数える）。範囲外は room.nickname_empty または room.nickname_too_long の 400 で返す。
    JoinRoomResponse:
      type: object
      required: [room, player]
      properties:
        room:
          $ref: "#/components/schemas/RoomSnapshot"
        player:
          $ref: "#/components/schemas/PlayerCredential"
    GetRoomResponse:
      type: object
      required: [room, playerId]
      properties:
        room:
          $ref: "#/components/schemas/RoomSnapshot"
        playerId:
          type: string
          description: 要求した本人のプレイヤー識別子。メンバー一覧のどれが自分かを決めるのに使う。
    StartGameResponse:
      type: object
      required: [room]
      properties:
        room:
          $ref: "#/components/schemas/RoomSnapshot"
    RoomNotification:
      type: object
      description: >-
        WebSocket で部屋のメンバーに送る通知。接続は GET /ws?roomCode=<部屋コード>&playerToken=<プレイヤートークン>
        で開く（ブラウザの WebSocket はヘッダーを付けられないのでクエリで渡す）。部屋が無い・期限切れなら 404、
        メンバーでなければ 403 で接続を拒否する。接続できると、直後に type が connected の通知でいまの部屋が届き、
        以後は部屋が変わるたびに届く。通知にプレイヤートークンは含まれない。
      required: [type, room]
      properties:
        type:
          type: string
          enum: [connected, member_joined, game_started]
          description: connected は接続直後のいまの部屋、member_joined はメンバーが加わった、game_started はゲームが始まった。
        room:
          $ref: "#/components/schemas/RoomSnapshot"
    RoomSnapshot:
      type: object
      required: [code, roundCount, hostPlayerId, status, members]
      properties:
        code:
          type: string
          description: 部屋コード。英大文字と数字 6 文字。
        roundCount:
          type: integer
        hostPlayerId:
          type: string
        status:
          type: string
          enum: [waiting, started]
          description: 進行状態。waiting は待機中、started はゲームが始まった後。
        members:
          type: array
          description: 開始前は部屋に入った順、開始後は描く順番。
          items:
            $ref: "#/components/schemas/RoomMember"
    RoomMember:
      type: object
      required: [playerId, nickname]
      properties:
        playerId:
          type: string
        nickname:
          type: string
    PlayerCredential:
      type: object
      description: 応答を受けた本人だけが持つ資格。
      required: [playerId, playerToken]
      properties:
        playerId:
          type: string
        playerToken:
          type: string
          description: 本人だけが持つ秘密。他のメンバーには配らない。
    ErrorResponse:
      type: object
      required: [code, detail]
      properties:
        code:
          type: string
          description: "失敗の種類。room.<snake_case>（room.nickname_empty, room.nickname_too_long, room.round_count_out_of_range, room.room_not_found, room.room_full, room.game_already_started, room.nickname_taken, room.not_member, room.not_host, room.not_enough_members など）、または request.invalid_json, request.invalid_body, server.internal_error。"
        detail:
          type: string